feat: mob AI, pathfinding, spawn, trees, redstone, safer shutdown
- Iron golem: multi-hostile targeting, arm-raise attack, no water pathing
- Skeleton bow equip/draw/shoot; clear target on death
- Zombie flee-sun; warden basic hearing/melee; path step-down fixes
- Creature natural spawn surface Y + raw brightness; spruce/pine canopy
- Piston NBT and immovable blocks; redstone wire client notify
- Ctrl+C shutdown timeouts and second-signal force exit
- Revert experimental keep-alive changes to stock behavior

+++ pumpkin/src/block/blocks/cake.rs
@@ -143,7 +143,7 @@ impl BlockBehaviour for CakeBlock {
                             BlockFlags::NOTIFY_ALL,
                         )
                         .await;
-                    let seed: f64 = rng().random();
+                    let seed: i64 = rng().random();
                     args.player
                         .play_sound(
                             Sound::BlockCakeAddCandle as u16,
